+++ esbuild.config.mjs
@@ -1,4 +1,4 @@
-import builtins from "builtin-modules";
+import { builtinModules } from "node:module";
 import esbuild from "esbuild";
 
 const prod = process.argv[2] === "production";
@@ -23,7 +23,8 @@ const context = await esbuild.context({
     "@lezer/common",
     "@lezer/highlight",
     "@lezer/lr",
-    ...builtins
+    ...builtinModules,
+    ...builtinModules.map((moduleName) => `node:${moduleName}`)
   ],
   format: "cjs",
   target: "es2018",